+++ src/lib.rs
@@ -10,7 +10,11 @@ extern crate termion;
 extern crate unicode_segmentation;
 extern crate unicode_width;
 
+#[deny(missing_docs)]
 pub mod base;
+#[deny(missing_docs)]
 pub mod container;
+#[deny(missing_docs)]
 pub mod input;
+#[deny(missing_docs)]
 pub mod widget;